# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dance Analysis API",
    "description": "Video dance analysis with pose estimation via microservices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sessions/{session_id}/access": {
      "get": {
        "tags": [
          "session-access"
        ],
        "summary": "List Session Access",
        "description": "List all users with access to a session.\n\nRequires: user is the session owner.",
        "operationId": "list_session_access_api_v1_sessions__session_id__access_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionAccessResponse"
                  },
                  "title": "Response List Session Access Api V1 Sessions  Session Id  Access Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/access/users": {
      "post": {
        "tags": [
          "session-access"
        ],
        "summary": "Grant Direct Access",
        "description": "Grant direct access to a user for a session.\n\nCreates both the SessionAccess record and tracks the origin.\n\nRequires: user is the session owner.",
        "operationId": "grant_direct_access_api_v1_sessions__session_id__access_users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionAccessCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/access/users/{user_id}": {
      "delete": {
        "tags": [
          "session-access"
        ],
        "summary": "Revoke Direct Access",
        "description": "Revoke direct access for a user on a session.\n\nRemoves the direct access origin. If no other origins remain,\nremoves the user's access entirely.\n\nRequires: user is the session owner.",
        "operationId": "revoke_direct_access_api_v1_sessions__session_id__access_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/access/origins": {
      "get": {
        "tags": [
          "session-access"
        ],
        "summary": "List Access Origins",
        "description": "List all access origins for a session.\n\nShows how each user gained access (direct or via group).\n\nRequires: user has access to the session.",
        "operationId": "list_access_origins_api_v1_sessions__session_id__access_origins_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionAccessOriginResponse"
                  },
                  "title": "Response List Access Origins Api V1 Sessions  Session Id  Access Origins Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/access/users/{user_id}/origins": {
      "get": {
        "tags": [
          "session-access"
        ],
        "summary": "List User Origins",
        "description": "List all origins for a specific user on a session.\n\nShows all ways the user has access (direct, multiple groups, etc.).\n\nRequires: user has access to the session.",
        "operationId": "list_user_origins_api_v1_sessions__session_id__access_users__user_id__origins_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionAccessOriginResponse"
                  },
                  "title": "Response List User Origins Api V1 Sessions  Session Id  Access Users  User Id  Origins Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/groups": {
      "get": {
        "tags": [
          "session-access"
        ],
        "summary": "List Session Groups",
        "description": "List all groups linked to a session.\n\nShows which groups have access to this session.\n\nRequires: user has access to the session.",
        "operationId": "list_session_groups_api_v1_sessions__session_id__groups_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionGroupLinkResponse"
                  },
                  "title": "Response List Session Groups Api V1 Sessions  Session Id  Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "session-access"
        ],
        "summary": "Link Group To Session",
        "description": "Link a group to a session.\n\nAll active group members will gain access to the session at the specified role.\nThis is atomic: group link and access records are created together.\n\nRequires: user is the session owner.",
        "operationId": "link_group_to_session_api_v1_sessions__session_id__groups_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionGroupLinkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionGroupLinkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/groups/{group_id}": {
      "delete": {
        "tags": [
          "session-access"
        ],
        "summary": "Unlink Group From Session",
        "description": "Unlink a group from a session.\n\nAll group-derived access will be removed for group members.\nIf a user has no other access origins, their access is revoked entirely.\n\nRequires: user is the session owner.",
        "operationId": "unlink_group_from_session_api_v1_sessions__session_id__groups__group_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/access/groups/{group_id}/users/{user_id}": {
      "delete": {
        "tags": [
          "session-access"
        ],
        "summary": "Revoke Group Member Access",
        "description": "Revoke group-derived access for a specific user.\n\nRemoves the group access origin for this user. If no other origins remain,\nremoves the user's access entirely.\n\nRequires: user is the session owner.",
        "operationId": "revoke_group_member_access_api_v1_sessions__session_id__access_groups__group_id__users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Group Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionAccessCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Access role: admin, editor, viewer",
            "default": "admin"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "SessionAccessCreate",
        "description": "Request to create/grant session access."
      },
      "SessionAccessOriginResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "user_id",
          "source_type",
          "created_at"
        ],
        "title": "SessionAccessOriginResponse",
        "description": "Response for session access origin."
      },
      "SessionAccessResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "user_id",
          "role",
          "created_at"
        ],
        "title": "SessionAccessResponse",
        "description": "Response for session access record."
      },
      "SessionGroupLinkCreate": {
        "properties": {
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Default role for group members: admin, editor, viewer",
            "default": "admin"
          }
        },
        "type": "object",
        "required": [
          "group_id"
        ],
        "title": "SessionGroupLinkCreate",
        "description": "Request to link a group to a session."
      },
      "SessionGroupLinkResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "group_id",
          "created_at"
        ],
        "title": "SessionGroupLinkResponse",
        "description": "Response for session-group link."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}